Upgrade urfave cli to v3 alpha for persistent command support

+++ internal/commands/constraints.go
@@ -10,14 +10,14 @@ func requireValidEnumValue(value string, allowedValues []string, enumName string
 	if !slices.Contains(allowedValues, value) {
 		return fmt.Errorf("flag %s must be one of [%s]", enumName, strings.Join(allowedTransportTypes, ", "))
 	}
-	
+
 	return nil
 }
 
-func requireIntInRange(value int, lower int, upper int, flagName string) error {
+func requireIntInRange(value int64, lower int64, upper int64, flagName string) error {
 	if value < lower || value > upper {
 		return fmt.Errorf("flag %s must be between %d and %d", flagName, lower, upper)
 	}
-	
+
 	return nil
 }
